FilterDropdown (#1688)
styling & functionality (work in progress)
