+++ components/filter-dropdown/index.css
@@ -1,5 +1,63 @@
 @import './components/NestedCheckbox/index.css';
 
 .filter-dropdown {
-  /* TODO */
+  position: relative;
+
+  &__select {
+    border: 1px solid rgba(var(--col-blue-mid-2));
+    border-radius: var(--br);
+    background-color: rgb(var(--col-background-primary-white));
+    display: flex;
+    align-items: stretch;
+
+    .filter-dropdown.is-opened & {
+      border-radius: var(--br) var(--br) 0 0;
+    }
+  }
+
+  &__label {
+    font-size: 1.125rem;
+    padding: .5625rem 1rem;
+    font-weight: var(--fw-bold);
+    line-height: 1.25rem;
+    flex: 1 1;
+    user-select: none;
+  }
+
+  &__icon {
+    flex: 0 0 40px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border-left: 1px solid rgba(var(--col-blue-mid-2));
+
+    svg {
+      margin-left: 0;
+    }
+  }
+
+  &__options-label {
+    padding-bottom: 1.5rem;
+  }
+
+  &__body {
+    position: absolute;
+    top: 100%;
+    left: 0;
+    right: 0;
+    padding: 1.5rem;
+    background-color: rgb(var(--col-background-primary-white));
+    display: none;
+    border: 1px solid rgba(var(--col-blue-mid-2));
+    border-top: none;
+    border-radius: 0 0 var(--br) var(--br);
+
+    .filter-dropdown.is-opened & {
+      display: block;
+    }
+  }
+
+  &__actions {
+    padding-top: 1.5rem;
+  }
 }
